+++ test/unit/network/PacketBuffer_unit_tests.cpp
@@ -18,42 +18,42 @@ namespace
 TEST(PacketBuffer_unit_tests, Append_Bool_AdjustsSize)
 {
     PacketBuffer buffer;
-    buffer.size = WriteBuffer(buffer.data, 0u, vBool);
+    buffer.size = static_cast<uint32_t>(WriteBuffer(buffer.data, 0u, vBool));
     EXPECT_EQ(buffer.size, sizeof(vBool));
 }
 
 TEST(PacketBuffer_unit_tests, Append_Uint8_AdjustsSize)
 {
     PacketBuffer buffer;
-    buffer.size = WriteBuffer(buffer.data, 0u, vUint8);
+    buffer.size = static_cast<uint32_t>(WriteBuffer(buffer.data, 0u, vUint8));
     EXPECT_EQ(buffer.size, sizeof(vUint8));
 }
 
 TEST(PacketBuffer_unit_tests, Append_Uint32_AdjustsSize)
 {
     PacketBuffer buffer;
-    buffer.size = WriteBuffer(buffer.data, 0u, vUint32);
+    buffer.size = static_cast<uint32_t>(WriteBuffer(buffer.data, 0u, vUint32));
     EXPECT_EQ(buffer.size, sizeof(vUint32));
 }
 
 TEST(PacketBuffer_unit_tests, Append_Float_AdjustsSize)
 {
     PacketBuffer buffer;
-    buffer.size = WriteBuffer(buffer.data, 0u, vFloat);
+    buffer.size = static_cast<uint32_t>(WriteBuffer(buffer.data, 0u, vFloat));
     EXPECT_EQ(buffer.size, sizeof(vFloat));
 }
 
 TEST(PacketBuffer_unit_tests, Append_String_AdjustsSize)
 {
     PacketBuffer buffer;
-    buffer.size = WriteBuffer(buffer.data, 0u, vCharArray);
+    buffer.size = static_cast<uint32_t>(WriteBuffer(buffer.data, 0u, vCharArray));
     EXPECT_EQ(buffer.size, std::strlen(vCharArray) + 1u);
 }
 
 TEST(PacketBuffer_unit_tests, ReadBuffer_HeterogeneousData_OutputMatchesInput)
 {
     PacketBuffer buffer;
-    buffer.size = WriteBuffer(buffer.data, 0u, vFloat, vUint32, vCharArray, vUint8, vBool);
+    buffer.size = static_cast<uint32_t>(WriteBuffer(buffer.data, 0u, vFloat, vUint32, vCharArray, vUint8, vBool));
 
     float actualFloat;
     uint32_t actualUint32;
